Modified 2022 Impact Factor for 2 journals
Two journals had "<0.1" and "NA" values, then putting them in the top25 list when sorting in descending order. Therefore they ended up in the top 25 list, and were removed with the aplication of the 100papers threshold.
So I gave them "0" as 2022 Impact Factor, they are out of the Top25 and 2 other journals take their place. And these two other journals do remain in the list even after the 100 papers threshold. So all the paper should have been made with a top22 journals instead of the 20 journals.